ref: keep pure functions for game-core together make it cleaner to read
